Basic Edition: - countdown timer was accidentally activated
All:
- the keyboard control for osae komi did always trigger toketa and did not switch fighters

+++ base/tournamentmodel.cpp
@@ -240,7 +240,7 @@ QVariant TournamentModel::headerData(
 		if (Qt::Vertical == orientation)
 			return QString::number(section + 1);
 
-		if (section < sizeof(m_HeaderData))
+		if (section < (int)sizeof(m_HeaderData))
 			return m_HeaderData[section];
 	}
